Add generated methods and constants
The generator has been enhanced to emit methods and constants. These
are emitted in a file Apis.ahk in the directory for the relevant
namespace, and are named after the namespace. This way, calling these
methods looks reasonable:
	Gdi.GetDC(0)

The changes to the generator also improved the type documentation for
pointers to pseudo-primitive structs.