Add Methods and Constants (#4)
* Add generated methods and constants

The generator has been enhanced to emit methods and constants. These
are emitted in a file Apis.ahk in the directory for the relevant
namespace, and are named after the namespace. This way, calling these
methods looks reasonable:
	Gdi.GetDC(0)

The changes to the generator also improved the type documentation for
pointers to pseudo-primitive structs.

* Update example to use generated methods

* Full name on the license

* Update README for methods

* Add example to list display devices

* Fix Layouts for Sequential structs w/nonzero packing size

Fixed an issue[1] in the generator where we incorrectly assumed that if
a struct's layout had a nonzero packing size, we should use it. However
there are a few structs that have a LayoutKind of sequential, but which
still have nonzero packing sizes.

[1] https://github.com/holy-tao/AhkWin32Structs-Generator/commit/462dea6604d35e9c95ff36784a19f35be910d359