Allow AHK strings in PSTR* and PWSTR* method arguments
For method arguments which are pointers to strings, we now allow native
AutoHotkey strings (or variables containing strings) to be passed. In
this case, we will obtain a pointer to the internal string buffer with
StrPtr and pass that to DllCall. String arguments are still ultimately
passed to DllCall as pointers.

This allows callers to provide a pointer, buffer, buffer-like object,
or a string literal or constructed string.