Update with generator fixes and Api Improvements
Api parameters which are pointers to primitives are now passed using
star-syntax [1] like int*. This also allows callers to pass VarRefs in
cases where the function they are calling will modify the variable.

Fixed an issue where the documented type of a Pointer generated from a
pseudo-primitive was incorrect. In English, an HWND value showed up in
the documentation as:
	Pointer<HWND>
When it should have been (and now is):
	Pointer<Void>
Because HWND is represented in the metadata as a wrapper struct for a
value of type void*.

[1] https://www.autohotkey.com/docs/v2/lib/DllCall.htm#asterisk