Respect PreserveSig and related attributes
Changes from generator update [1]. Methods that return HRESULT values
will now throw OSErrors if the result is anything other than E_SUCCESS,
unless they are decorated with one of the following attributes:
	- [CanReturnMultipleSuccessValues]
    - [CanReturnErrorsAsSuccess]
In this case, a non-success HRESULT may not indicate an error. Callers
are responsible for reading the documentation and checking errors
manually.

HRESULT values are now documented as HRESULTs instead of Integers, but
under the hood they are still 32-bit Integers.

[1] https://github.com/holy-tao/AhkWin32Structs-Generator/commit/2d49ea2fbfe89f4c3b460e12c2f6d52aeefa6aab