Add a Guid wrapper struct
To properly work with COM, we'll need a convenient way to interact with
GUIDs both in struct form so we can obtain pointers to them, and in
string form, since most AHK native COM interop methods expect GUIDs as
strings. Hand-roll a new proxy object and add tests for it.