Add base class for COM interfaces
Add a new base class for generated COM interfaces. AHK's native COM
interop is too abstract for what we want to do here, the main issue is
that it only supports IDispatch, but it also handles reference counting
in odd ways. Add a base class for shared methods. In practice this will
mostly be a wrapper for a pointer.