Allow initialization from objects in __New
Allows users to pass in an object literal to populate struct members
from into __New. In this case, a new Buffer is allocated.

Doing this now because Handles will require frequent initialization
from values, and FromObject is already kind of clunky. They also need
to extend Win32Struct's default __New implementation, so FromObject,
which always calls __New with no arguments, wasn't a full solution to
the issue regardless.

FromObject is retained. Added tests for __New with objects, kept the
tests for FromObject.

Resolves: https://github.com/holy-tao/AhkWin32Projection/issues/21