Track embedded struct lineage
Structs can how take a parent parameter, which indicates that they are
embedded in another struct. This is important for handles because they
inherit ownership. If a struct has a parent, its pointer is calculated
as an offset from its parent's pointer. This requires a few changes to
initialization code in __New and _InitFromObject - we now need to care
a lot about the order in which struct properties are initialized.