Use typed handles
Handle types like HINSTANCE are now emitted as Win32Structs. We use the
[RAIIFree] and [InvalidHandleValue] attributes on these handles to add
additional functionality to them, most notably to free owned handles in
__Delete() when they fall out of scope.

This also introduces the concept of "ownership" for structures although
it has no effect other than that handles in structs inherit it from the
structs they're embedded in.

Apis that return handles return handle objects wrapping the actual ints
which are always script-owned. Apis that take handles as parameters now
accept and automatically dereference typed handles if passed.