Allow raw pointers and VarRefs primitive-pointer arguments
Pointers to primitives in method signatures (int*, float*) were always
passed to DllCall with an asterisk type, to allow users to pass VarRefs
and get responses back. But, as the issue [1] notes, it was impossible
to ass raw pointers or NULL as arguments to methods.

Only use asterisk types if passed a VarRef - otherwise, pass the param
to DllCall as a raw "ptr"

[1] https://github.com/holy-tao/AhkWin32Structs-Generator/issues/18