Add interface properties for  methods
Com interface methods with the [SpecialName] attribute are now used to
generate properties on the interfaces. For example, if an interface has
the following methods:
	get_Name()
	put_Name(value)
We will generate a "Name" property with a getter and setter that invoke
those methods. Properties can be read-only - if a "put_*" method does
not exist, no setter will be generated.