Don't generate invalid free functions
Previously, handles with [RAIIFree] or [FreeWith] attributes would
blindly call the function named in the attribute, assuming that it both
took only one argument. In a few cases, this turns out to be untrue,
and the automatic freeing logic would cause errors to be thrown.

Now, we only generate free functions for handles if we can tell for
certain that it's safe to call the function with the handle as its only
argument.